Cambios para la carga de producto

+++ IJM/src/main/java/com/IJM/dao/ProductDao.java
@@ -12,7 +12,7 @@ public interface ProductDao {
 
 	public List<Product> findAll();
 
-	public Product findById(int id);
+	public Product findById(Long id);
 	
 	public Product findByCode(String code);
